+++ docs/changes.rst
@@ -26,6 +26,8 @@ To be released.
   :meth:`~sqlalchemy_imageattach.entity.BaseImageSet.from_blob()` can take
   ``extra_args``/``extra_kwargs`` to be passed to entity model's constructor.
   [:issue:`32`, :issue:`33` by Vahid]
+- Added :const:`sqlalchemy_imageattach.version.SQLA_COMPAT_VERSION` and
+  :const:`sqlalchemy_imageattach.version.SQLA_COMPAT_VERSION_INFO` constants.
 
 
 Version 0.9.0